Align projects list with mirror-needing projects
If projects.txt is going to be the basis for select-mirror in the
gate, we should also list projects in projects.txt that need to be
constrained by the mirror for their testing. In order to do that, we
need to not try to unconditionally update setup.py, since devstack and
devstack-gate do not have it.

openstack/requirements itself also does not use the mirror, so remove
it from the list.

+++ update.py
@@ -149,10 +149,13 @@ def _copy_requires(suffix, dest_dir):
 
 
 def _write_setup_py(dest_path):
-    print("Syncing setup.py")
     target_setup_py = os.path.join(dest_path, 'setup.py')
+    # If it doesn't have a setup.py, then we don't want to update it
+    if not os.path.exists(target_setup_py):
+        return
     has_pbr = 'pbr' in open(target_setup_py, 'r').read()
     if has_pbr:
+        print("Syncing setup.py")
         # We only want to sync things that are up to date with pbr mechanics
         with open(target_setup_py, 'w') as setup_file:
             setup_file.write(_setup_py_text)
